+++ builder_pkgs/mockito/lib/src/builder.dart
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
@@ -38,67 +39,75 @@ class MockBuilder implements Builder {
     final entryLib = await buildStep.inputLibrary;
     final resolver = buildStep.resolver;
 
-    final mockLibrary = buildStep.inputId.changeExtension('.mocks.dart');
+    final mockLibraryAsset = buildStep.inputId.changeExtension('.mocks.dart');
 
-    final mockClasses = <Class>[];
-
-    for (final element in entryLib.topLevelElements) {
-      final annotation = element.metadata.firstWhere(
-          (annotation) =>
-              annotation.element is ConstructorElement &&
-              annotation.element.enclosingElement.name == 'GenerateMocks',
-          orElse: () => null);
-      if (annotation == null) continue;
-      final generateMocksValue = annotation.computeConstantValue();
-      // TODO(srawlins): handle `generateMocksValue == null`?
-      final classesToMock = generateMocksValue.getField('classes');
-      if (classesToMock.isNull) {
-        throw InvalidMockitoAnnotationException(
-            'The "classes" argument has unknown types');
-      }
-      for (final classToMock in classesToMock.toListValue()) {
-        final dartTypeToMock = classToMock.toTypeValue();
-        // TODO(srawlins): Import the library which declares [dartTypeToMock].
-        // TODO(srawlins): Import all supporting libraries, used in type
-        // signatures.
-        if (dartTypeToMock == null) {
+    final mockLibrary = Library((lBuilder) {
+      for (final element in entryLib.topLevelElements) {
+        final annotation = element.metadata.firstWhere(
+            (annotation) =>
+                annotation.element is ConstructorElement &&
+                annotation.element.enclosingElement.name == 'GenerateMocks',
+            orElse: () => null);
+        if (annotation == null) continue;
+        final generateMocksValue = annotation.computeConstantValue();
+        // TODO(srawlins): handle `generateMocksValue == null`?
+        final classesToMock = generateMocksValue.getField('classes');
+        if (classesToMock.isNull) {
           throw InvalidMockitoAnnotationException(
-              'The "classes" argument includes a non-type: $classToMock');
+              'The "classes" argument has unknown types');
         }
 
-        final elementToMock = dartTypeToMock.element;
-        if (elementToMock is ClassElement) {
-          if (elementToMock.isEnum) {
-            throw InvalidMockitoAnnotationException(
-                'The "classes" argument includes an enum: '
-                '${elementToMock.displayName}');
-          }
-          // TODO(srawlins): Catch when someone tries to generate mocks for an
-          // un-subtypable class, like bool, String, FutureOr, etc.
-          mockClasses.add(_buildCodeForClass(dartTypeToMock, elementToMock));
-        } else if (elementToMock is GenericFunctionTypeElement &&
-            elementToMock.enclosingElement is FunctionTypeAliasElement) {
-          throw InvalidMockitoAnnotationException(
-              'The "classes" argument includes a typedef: '
-              '${elementToMock.enclosingElement.displayName}');
-        } else {
-          throw InvalidMockitoAnnotationException(
-              'The "classes" argument includes a non-class: '
-              '${elementToMock.displayName}');
-        }
+        _buildMockClasses(classesToMock.toListValue(), lBuilder);
       }
-    }
+    });
 
-    if (mockClasses.isEmpty) {
+    if (mockLibrary.body.isEmpty) {
       // Nothing to mock here!
       return;
     }
 
-    final emitter = DartEmitter();
-    final mockLibraryContent = DartFormatter()
-        .format(mockClasses.map((c) => c.accept(emitter)).join('\n'));
+    final emitter = DartEmitter.scoped();
+    final mockLibraryContent =
+        DartFormatter().format(mockLibrary.accept(emitter).toString());
 
-    await buildStep.writeAsString(mockLibrary, mockLibraryContent);
+    await buildStep.writeAsString(mockLibraryAsset, mockLibraryContent);
+  }
+
+  /// Build mock classes for [classesToMock], a list of classes obtained from a
+  /// `@GenerateMocks` annotation.
+  void _buildMockClasses(
+      List<DartObject> classesToMock, LibraryBuilder lBuilder) {
+    for (final classToMock in classesToMock) {
+      final dartTypeToMock = classToMock.toTypeValue();
+      // TODO(srawlins): Import the library which declares [dartTypeToMock].
+      // TODO(srawlins): Import all supporting libraries, used in type
+      // signatures.
+      if (dartTypeToMock == null) {
+        throw InvalidMockitoAnnotationException(
+            'The "classes" argument includes a non-type: $classToMock');
+      }
+
+      final elementToMock = dartTypeToMock.element;
+      if (elementToMock is ClassElement) {
+        if (elementToMock.isEnum) {
+          throw InvalidMockitoAnnotationException(
+              'The "classes" argument includes an enum: '
+              '${elementToMock.displayName}');
+        }
+        // TODO(srawlins): Catch when someone tries to generate mocks for an
+        // un-subtypable class, like bool, String, FutureOr, etc.
+        lBuilder.body.add(_buildCodeForClass(dartTypeToMock, elementToMock));
+      } else if (elementToMock is GenericFunctionTypeElement &&
+          elementToMock.enclosingElement is FunctionTypeAliasElement) {
+        throw InvalidMockitoAnnotationException(
+            'The "classes" argument includes a typedef: '
+            '${elementToMock.enclosingElement.displayName}');
+      } else {
+        throw InvalidMockitoAnnotationException(
+            'The "classes" argument includes a non-class: '
+            '${elementToMock.displayName}');
+      }
+    }
   }
 
   Class _buildCodeForClass(DartType dartType, ClassElement classToMock) {
@@ -107,7 +116,8 @@ class MockBuilder implements Builder {
     return Class((cBuilder) {
       cBuilder
         ..name = 'Mock$className'
-        ..extend = refer('Mock')
+        ..extend = refer('Mock', 'package:mockito/mockito.dart')
+        // TODO(srawlins): Add URI of dartType.
         ..implements.add(refer(className))
         ..docs.add('/// A class which mocks [$className].')
         ..docs.add('///')
@@ -125,14 +135,39 @@ class MockBuilder implements Builder {
         }
       }
       for (final method in classToMock.methods) {
-        if (method.parameters.isEmpty || method.isPrivate || method.isStatic) {
+        if (method.isPrivate || method.isStatic) {
           continue;
         }
-        cBuilder.methods.add(
-            Method((mBuilder) => _buildOverridingMethod(mBuilder, method)));
+        if (_returnTypeIsNonNullable(method) ||
+            _hasNonNullableParameter(method)) {
+          cBuilder.methods.add(
+              Method((mBuilder) => _buildOverridingMethod(mBuilder, method)));
+        }
       }
     });
   }
+
+  // TODO(srawlins): Update this logic to correctly handle non-nullable return
+  // types. Right now this logic does not seem to be available on DartType.
+  bool _returnTypeIsNonNullable(MethodElement method) {
+    var type = method.returnType;
+    if (type.isDynamic || type.isVoid) return false;
+    if (method.isAsynchronous && type.isDartAsyncFuture ||
+        type.isDartAsyncFutureOr) {
+      var typeArgument = (type as InterfaceType).typeArguments.first;
+      if (typeArgument.isDynamic || typeArgument.isVoid) {
+        // An asynchronous method which returns `Future<void>`, for example,
+        // does not need a dummy return value.
+        return false;
+      }
+    }
+    return true;
+  }
+
+  // TODO(srawlins): Update this logic to correctly handle non-nullable return
+  // types. Right now this logic does not seem to be available on DartType.
+  bool _hasNonNullableParameter(MethodElement method) =>
+      method.parameters.isNotEmpty;
 
   /// Build a method which overrides [method], with all non-nullable
   /// parameter types widened to be nullable.
@@ -149,6 +184,10 @@ class MockBuilder implements Builder {
     builder
       ..name = method.displayName
       ..returns = refer(method.returnType.displayName);
+
+    if (method.typeParameters != null && method.typeParameters.isNotEmpty) {
+      builder.types.addAll(method.typeParameters.map((p) => refer(p.name)));
+    }
 
     if (method.isAsynchronous) {
       builder.modifier =
@@ -180,16 +219,52 @@ class MockBuilder implements Builder {
     // type checks.
     // TODO(srawlins): Handle getter invocations with `Invocation.getter`,
     // and operators???
-    // TODO(srawlins): Handle generic methods with `Invocation.genericMethod`.
     final invocation = refer('Invocation').property('method').call([
       refer('#${method.displayName}'),
       literalList(invocationPositionalArgs),
       if (invocationNamedArgs.isNotEmpty) literalMap(invocationNamedArgs),
     ]);
+    final noSuchMethodArgs = <Expression>[invocation];
+    if (_returnTypeIsNonNullable(method)) {
+      final dummyReturnValue = _dummyValue(method.returnType);
+      noSuchMethodArgs.add(dummyReturnValue);
+    }
     final returnNoSuchMethod =
-        refer('super').property('noSuchMethod').call([invocation]);
+        refer('super').property('noSuchMethod').call(noSuchMethodArgs);
 
     builder.body = returnNoSuchMethod.code;
+  }
+
+  Expression _dummyValue(DartType type) {
+    if (type.isDartCoreBool) {
+      return literalFalse;
+    } else if (type.isDartCoreDouble) {
+      return literalNum(0.0);
+    } else if (type.isDartAsyncFuture || type.isDartAsyncFutureOr) {
+      var typeArgument = (type as InterfaceType).typeArguments.first;
+      return refer('Future')
+          .property('value')
+          .call([_dummyValue(typeArgument)]);
+    } else if (type.isDartCoreInt) {
+      return literalNum(0);
+    } else if (type.isDartCoreList) {
+      return literalList([]);
+    } else if (type.isDartCoreMap) {
+      return literalMap({});
+    } else if (type.isDartCoreNum) {
+      return literalNum(0);
+    } else if (type.isDartCoreSet) {
+      // This is perhaps a dangerous hack. The code, `{}`, is parsed as a Set
+      // literal if it is used in a context which explicitly expects a Set.
+      return literalMap({});
+    } else if (type.isDartCoreString) {
+      return literalString('');
+    } else {
+      // TODO(srawlins): Returning null for now, but really this should only
+      // ever get to a state where we have to make a Fake class which implements
+      // the type, and return a no-op constructor call to that Fake class here.
+      return literalNull;
+    }
   }
 
   /// Returns a [Parameter] which matches [parameter].
@@ -197,6 +272,7 @@ class MockBuilder implements Builder {
       Parameter((pBuilder) {
         pBuilder
           ..name = parameter.displayName
+          // TODO(srawlins): Add URI of `parameter.type`.
           ..type = refer(parameter.type.displayName);
         if (parameter.isNamed) pBuilder.named = true;
         if (parameter.defaultValueCode != null) {